# market-manager/cmd/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Market Manager API Server",
        "title": "Market Manager Api",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:10012",
    "basePath": "/api/v1/",
    "paths": {
        "/api/v1/cpod/job": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/cpod/job/result": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "cpod job info",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/common.CpodJobInfo"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Err"
                        }
                    }
                }
            }
        },
        "/api/v1/cpod/resource": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "description": "cpod resource info",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/common.CpodResourceInfos"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/common.Err"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.CpodJobInfo": {
            "type": "object",
            "properties": {
                "cpod_job_id": {
                    "type": "string"
                },
                "job_id": {
                    "type": "string"
                },
                "job_url": {
                    "type": "string"
                },
                "state": {
                    "type": "integer"
                }
            }
        },
        "common.CpodResourceInfo": {
            "type": "object",
            "properties": {
                "cpod_id": {
                    "type": "string"
                },
                "gpu_free": {
                    "type": "number"
                },
                "gpu_total": {
                    "type": "number"
                },
                "gpu_used": {
                    "type": "number"
                },
                "group_id": {
                    "type": "string"
                }
            }
        },
        "common.CpodResourceInfos": {
            "type": "object",
            "properties": {
                "cpods": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/common.CpodResourceInfo"
                    }
                }
            }
        },
        "common.Err": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "业务编码",
                    "type": "integer"
                },
                "data": {
                    "description": "成功时返回的数据"
                },
                "msg": {
                    "description": "错误描述",
                    "type": "string"
                }
            }
        }
    }
}